[Bronze V] Title: 사파리월드, Time: 124 ms, Memory: 14132 KB -BaekjoonHub

+++ 백준/Bronze/2420. 사파리월드/사파리월드.java
@@ -9,8 +9,8 @@ class Main{
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
         StringBuilder sb = new StringBuilder();
         
-        long n = Integer.parseInt(st.nextToken());
-        long m = Integer.parseInt(st.nextToken());
+        long n = Long.parseLong(st.nextToken());
+        long m = Long.parseLong(st.nextToken());
         
         if(n-m<0){
             sb.append(-(n-m));
